feat: Add many luck stats & copy new league info
I had so many edits for the revamped luck index, and forgot I had this branch checked out. So I'm committing all of them. Additionally, this commit is mainly to update the default year to 2024, add a `created_date` column to the database & LeagueInfo objects, and allows users to copy over old leagues for the current year.
