+++ .gitignore
@@ -11,3 +11,6 @@ dist/*
 src/new espn.ipynb
 draft_history.csv
 src/.DS_Store
+league_info_backup.csv
+src/tmp.csv
+website.ipynb
