+++ fantasy_stats/models.py
@@ -8,6 +8,7 @@ class LeagueInfo(models.Model):
     swid = models.CharField(max_length=36)
     espn_s2 = models.CharField(max_length=500)
     league_name = models.CharField(max_length=50, default="<Name Missing>")
+    created_date = models.DateTimeField(auto_now_add=True)
 
     def __str__(self):
         return "{} ({})".format(self.league_id, self.league_year)
